Move maya-agent code to a separate directory
1. Why is this change necessary ?
- As exporter needs to be run separately, therefore it is causing some
  issues with maya-agent. Both serve different purpose. One more
  difficulty with maya-agent is its useability i.e, it is planned to
  run as daemonset but exporter doesn't fit into this use-case.
- This also makes maya-volume-exporter code to be more simple and manageable.

2. How does this change address the issue ?
- Now maya-volume-exporter will be the new binary that can be used for
  monitoring volumes.

3. How to verify this change ?
- Run any instance of volumes, get the IP of volume controller and pass
  as flag while running binary.
- For exp : `maya-volume-exporter -c=localhost:9501`

4. What side effects does this change have ?
- Some changes has been made in maya-agent's command.go
- Updated GNUmakefile
- Added buildscripts for maya-volume-exporter

5. Other details
improvement: #897
- Need to create a deployment file for maya-volume-exporter
