+++ cmd/maya-volume-exporter/app/command/command.go
@@ -1,6 +1,8 @@
-package exporter
+package command
 
 import (
+	goflag "flag"
+
 	"github.com/golang/glog"
 	"github.com/openebs/maya/pkg/util"
 	"github.com/spf13/cobra"
@@ -13,16 +15,21 @@ type VolumeExporterOptions struct {
 	ControllerAddress string
 }
 
+// AddListenAddressFlag is used to create flag to pass the listen address of exporter.
 func AddListenAddressFlag(cmd *cobra.Command, value *string) {
 	cmd.Flags().StringVarP(value, "listen.addr", "a", *value,
 		"Address on which to expose metrics and web interface.)")
 }
 
+// AddMetricsPathFlag is used to create flag to pass the listen path where volume
+// metrics are exposed.
 func AddMetricsPathFlag(cmd *cobra.Command, value *string) {
 	cmd.Flags().StringVarP(value, "listen.path", "m", *value,
 		"Path under which to expose metrics.")
 }
 
+// AddControllerAddressFlag is used to create flag to pass the Jiva volume
+// controllers IP.
 func AddControllerAddressFlag(cmd *cobra.Command, value *string) {
 	cmd.Flags().StringVarP(value, "controller.addr", "c", *value,
 		"Address of the Jiva volume controller.")
@@ -30,7 +37,7 @@ func AddControllerAddressFlag(cmd *cobra.Command, value *string) {
 
 // NewCmdVolumeExporter is used to create command monitoring and it initialize
 // monitoring flags also.
-func NewCmdVolumeExporter() *cobra.Command {
+func NewCmdVolumeExporter() (*cobra.Command, error) {
 	// create an instance of VolumeExporterOptions to initialize with default
 	// values for the flags.
 	options := VolumeExporterOptions{}
@@ -38,28 +45,29 @@ func NewCmdVolumeExporter() *cobra.Command {
 	options.ListenAddress = ":9500"
 	options.MetricsPath = "/metrics"
 	cmd := &cobra.Command{
-		Use:   "monitor",
 		Short: "Collect metrics from OpenEBS volumes",
-		Long: `  monitor command is used to start monitoring openebs volumes. It
-  start collecting metrics from the jiva controller at the endpoint
-  "/v1/stats" and push it to Prometheus Server`,
-		Example: `  maya-agent monitor -a=http://localhost:8001 -c=:9500 -m=/metrics`,
+		Long: `  maya-volume-exporter monitors openebs volumes and exporter the metrics.
+  It starts collecting metrics from the jiva volume controller at the endpoint
+  "/v1/stats" Prometheus Server can collect the metrics from maya-volume-exporter.`,
+		Example: `  maya-agent -a=http://localhost:8001 -c=:9500 -m=/metrics`,
 		Run: func(cmd *cobra.Command, args []string) {
 			util.CheckErr(Run(cmd, &options), util.Fatal)
 		},
 	}
 
+	cmd.Flags().AddGoFlagSet(goflag.CommandLine)
+	goflag.CommandLine.Parse([]string{})
 	AddControllerAddressFlag(cmd, &options.ControllerAddress)
 	AddListenAddressFlag(cmd, &options.ListenAddress)
 	AddMetricsPathFlag(cmd, &options.MetricsPath)
 
-	return cmd
+	return cmd, nil
 }
 
-// run used to process commands,args and call openebs exporter and it returns
+// Run used to process commands,args and call openebs exporter and it returns
 // nil on successful execution.
 func Run(cmd *cobra.Command, options *VolumeExporterOptions) error {
-	glog.Infof("Starting openebs-exporter ...")
+	glog.Infof("Starting maya-volume-exporter ...")
 	Entrypoint(options)
 	return nil
 }
